Push progress

+++ app/src/main/java/android/example/ub_durensawit/DbConn/DataSource/CartRepository.java
@@ -106,7 +106,7 @@ public class CartRepository {
         return cartDatabase.cartDao(). GetCartItemById(id);
     }
 
-    public LiveData<BigDecimal> getTotalHarga() {
+    public LiveData<Double> getTotalHarga() {
         return cartDatabase.cartDao().getTotal_Harga();
     }
